+++ tenpy/algorithms/purification.py
@@ -9,7 +9,7 @@ logger = logging.getLogger(__name__)
 from ..linalg import np_conserved as npc
 from . import tebd
 from .mps_common import VariationalApplyMPO, TwoSiteH
-from .truncation import svd_theta, TruncationError
+from ..linalg.truncation import svd_theta, TruncationError
 from .disentangler import get_disentangler
 
 __all__ = ['PurificationTwoSiteU', 'PurificationApplyMPO', 'PurificationTEBD', 'PurificationTEBD2']
