+++ tenpy/linalg/__init__.py
@@ -19,16 +19,18 @@ so you probably won't need to import `charges` directly.
     random_matrix
     sparse
     krylov_based
+    truncation
 
 """
 # Copyright (C) TeNPy Developers, GNU GPLv3
 
-from . import charges, np_conserved, krylov_based, random_matrix, sparse, svd_robust
+from . import charges, np_conserved, krylov_based, random_matrix, sparse, svd_robust, truncation
 from .charges import *
 from .np_conserved import *
 from .krylov_based import *
 from .random_matrix import *
 from .sparse import *
+from .truncation import *
 
 __all__ = ['charges', 'np_conserved', 'krylov_based', 'random_matrix', 'sparse', 'svd_robust',
            *charges.__all__,
@@ -36,6 +38,7 @@ __all__ = ['charges', 'np_conserved', 'krylov_based', 'random_matrix', 'sparse',
            *krylov_based.__all__,
            *random_matrix.__all__,
            *sparse.__all__,
+           *truncation.__all__,
            ]
 
 from ..tools import optimization
