+++ tests/test_tdvp.py
@@ -6,7 +6,7 @@ from tenpy.models.spins import SpinChain
 from tenpy.algorithms import tdvp
 from tenpy.algorithms import tebd
 from tenpy.networks.mps import MPS
-from tenpy.algorithms.truncation import TruncationError
+from tenpy.linalg.truncation import TruncationError
 
 
 @pytest.mark.slow
@@ -103,7 +103,7 @@ def test_fixes_339(L=8, chi=10, dt=0.5):
     err_max = np.max(tdvp2.trunc_err_list)
     wrong_err = expect_err + TruncationError(err_max, 1 - 2 * err_max)
 
-    
+
     if np.allclose(tdvp2.trunc_err.eps, wrong_err.eps):
         print('matches with wrong eps')
     if np.allclose(tdvp2.trunc_err.ov_err, wrong_err.ov_err):
@@ -113,4 +113,4 @@ def test_fixes_339(L=8, chi=10, dt=0.5):
     assert np.allclose(tdvp2.trunc_err.ov_err, expect_err.ov_err)
     assert not np.allclose(tdvp2.trunc_err.eps, wrong_err.eps)
     assert not np.allclose(tdvp2.trunc_err.ov_err, wrong_err.ov_err)
-    
+
